Created class Pearson
This class will contains the Pearson Correlation Score. As Joel Cogley
said in his article, the algorithm is:
"first finds the reviewers that reviewed both products. It then
computes the sums and the squared sums of the ratings for the two
products, and
computes the sum of the reviews of the products. Finally,
it uses these results to compute the Pearson Correlation Score."